1.增加:新版版本

+++ Hub/Hub/applications/Uart/DeviceUartClass/DeviceUartClass.h
@@ -52,7 +52,7 @@ typedef struct uartClass{
     void(*DeviceCtrlSingle)(device_t *, u8 , u8 );
     void(*DeviceChgType)(type_monitor_t *, u16 id, u8 type);       //发送设置端口type
     void (*LineCtrl)(line_t *line, u8 port, u8 state, u8 value);
-    void (*Line4Ctrl)(line_t *line, u8 *value);
+    void (*Line4Ctrl)(line_t *line, u16 *value);
     void(*AskDevice)(device_t , u16);                           //发送串口数据
     void(*AskSensor)(sensor_t , u16);                           //发送串口数据
     void(*AskLine)(line_t , u16);                               //发送串口数据
